% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_df_from_binary}
\alias{get_df_from_binary}
\title{Scan a binary TransCAD File}
\usage{
get_df_from_binary(bin_file, name, type, start, width, row_length)
}
\arguments{
\item{str}{bin_file string with the path to the .bin file.}

\item{CharacterVector}{name name of each variable.}

\item{CharacterVector}{type the information type for each variable.}

\item{NumericVector}{start the number of bytes in which the variable starts.}

\item{NumericVector}{end the number of bytes in which the variable ends.}

\item{int}{row_length the number of bytes in a row.}
}
\value{
characters in each element of the vector
}
\description{
This is a C++ implementation written by Amar Sarvepalli and adapted for Rcpp
by Greg Macfarlane.
}
\details{
All input parameter values are available in the \code{.DCB} file
  read by \link{read_dcb}. This is an internal function and should not
  normally be used independent of \link{read_tcad_bin}.
}

