% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_width}
\alias{get_width}
\title{get field width}
\usage{
get_width(s)
}
\arguments{
\item{s}{character field}
}
\value{
element_width width of the column
}
\description{
get field width
}
\details{
Method to get column width
}

