% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getColumns}
\alias{getColumns}
\title{get columns}
\usage{
getColumns(df)
}
\arguments{
\item{df}{DataFrame object}
}
\value{
ncols number of columns
}
\description{
get columns
}
\details{
To get number of dataframe columns
}

