% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNames}
\alias{getNames}
\title{get field names}
\usage{
getNames(df)
}
\arguments{
\item{df}{DataFrame object}
}
\value{
names field names
}
\description{
get field names
}
\details{
Method to extract dataframe column names
}

