% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRows}
\alias{getRows}
\title{get rows}
\usage{
getRows(df)
}
\arguments{
\item{df}{DataFrame object}
}
\value{
nrows number of rows
}
\description{
get rows
}
\details{
To get number of dataframe rows
}

