% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{read_assignment}
\alias{read_assignment}
\title{Read in TransCAD assignment file}
\usage{
read_assignment(file, field = NULL)
}
\arguments{
\item{file}{Path to assignment binary file.}

\item{field}{The name of the field to collect; if \code{NULL} (default), then
will return all fields.}
}
\description{
TransCAD returns fairly uniform assignment results containing
  times, speeds, volumes, etc. by direction. This function returns the flow
  information as a tidy data frame
}

