% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_tcad.R
\name{read_dcb}
\alias{read_dcb}
\title{Read the TransCAD dictionary file}
\usage{
read_dcb(dcb_file)
}
\arguments{
\item{dcb_file}{A TransCAD dictionary file containing information on the
layout of the binary data file and its attribute properties.}
}
\value{
A \code{data_frame} of the attributes including type, description,
and layout of the rows in the binary data.
}
\description{
Read the TransCAD \code{.DCB} dictionary file. This is a helper
function and is not normally called by the user directly.
}

