% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_tcad.R
\name{read_tcad_bin}
\alias{read_tcad_bin}
\title{Read a TransCAD binary file}
\usage{
read_tcad_bin(dir, file_name)
}
\arguments{
\item{dir}{The path to the directory where the files are located}

\item{file_name}{The name of the binary file, e.g. \code{foo.bin}}
}
\value{
A list with two elements: \code{description} contains any data labels
on the dictionary file; \code{data} is a \code{tbl_df(data_frame)} object with
the table.
}
\description{
Read a TransCAD \code{.bin} binary data file into the workspace.
}

