% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcad.R
\name{trim}
\alias{trim}
\title{Trim all duplicate spaces from a character string.}
\usage{
trim(x)
}
\arguments{
\item{x}{a character string.}
}
\value{
a character string with extraneous spaces removed.
}
\description{
Trim all duplicate spaces from a character string.
}
