% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tcad.R
\name{write_bin}
\alias{write_bin}
\title{Write a TransCAD binary file}
\usage{
write_bin(df, file)
}
\arguments{
\item{df}{workspace dataframe object}

\item{file}{The path and name of the binary data file, e.g.
\code{data/foo.bin}.}
}
\value{
The \code{DCB} dictionary file will be written to the same
  folder as the \code{.bin} file.
}
\description{
Write a TransCAD \code{.bin} binary data file from the workspace.
}
\details{
This version doesn't work well with NA's in the dataframe.
}

