% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_dcb}
\alias{write_dcb}
\title{write dictionary file}
\usage{
write_dcb(df, file_name)
}
\arguments{
\item{df}{DataFrame object}

\item{file_name}{name of the file}
}
\description{
write dictionary file
}
\details{
internal function dcb info
}

