% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_df_to_binary}
\alias{write_df_to_binary}
\title{write binary file}
\usage{
write_df_to_binary(df, file_name, field_types)
}
\arguments{
\item{df}{DataFrame object}

\item{file_name}{output filename}

\item{field_types}{datatypes}
}
\value{
new_dcb_info_df prints dictionary file to console
}
\description{
write binary file
}
\details{
Method to export data from dataframe to binary file.
}

