% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_df_from_binary}
\alias{get_df_from_binary}
\title{Scan a binary TransCAD File}
\usage{
get_df_from_binary(bin_file, name, type, start, width, row_length)
}
\arguments{
\item{bin_file}{string with the path to the .bin file.}

\item{name}{character vector of each variable.}

\item{type}{CharacterVector with the information type for each variable.}

\item{start}{NumericVector with the number of bytes in which the variable
starts.}

\item{width}{NumericVector with the number of bytes given for a variable.}

\item{row_length}{int showing the number of bytes in a row.}
}
\value{
characters in each element of the vector
}
\description{
This is a C++ implementation written by Amar Sarvepalli and adapted for Rcpp
by Greg Macfarlane.
}
\details{
All input parameter values are available in the \code{.DCB} file 
  read by \link{read_dcb}. This is an internal function and should not
  normally be used independent of \link{read_tcad_bin}.
}

