% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDims}
\alias{getDims}
\title{Get the dimensions of a data frame}
\arguments{
\item{df}{DataFrame object}
}
\value{
Array with the number of rows and columns (r x c)
}
\description{
Get the dimensions of a data frame
}
\details{
To get number of dataframe dimensions
}
