% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDims}
\alias{getDims}
\title{get dimensions}
\usage{
getDims(df)
}
\arguments{
\item{df}{DataFrame object}
}
\value{
nrowcol array of rows and columns (r x c)
}
\details{
To get number of dataframe dimensions
}

