% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{name_split}
\alias{name_split}
\title{parse line by comma deliminated}
\usage{
name_split(line, delim)
}
\arguments{
\item{line}{string line separated by delimiter}

\item{delim}{delimter can be anything}
}
\value{
fields elements of string line
}
\details{
splits line by ","
}

