% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{name_split}
\alias{name_split}
\title{Parse comma-delimited names line}
\arguments{
\item{line}{string line separated by delimiter}

\item{delim}{delimter can be anything}
}
\value{
fields elements of string line
}
\description{
Parse comma-delimited names line
}
\details{
splits line by ","
}

