% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tcad.R
\name{write_bin}
\alias{write_bin}
\title{Write a TransCAD binary file}
\usage{
write_bin(df, file_name)
}
\arguments{
\item{df}{workspace dataframe object}

\item{file_name}{The path and name of the binary data file, e.g.
\code{data/foo.bin}.}
}
\value{
The \code{DCB} dictionary file will be writern to the same
  folder. Prints dcb file contents to console.Need to make it a user defined 
  option.
}
\description{
Write a TransCAD \code{.bin} binary data file fromt the workspace.
}
\details{
If any factors exist in the data frame, then those fields are writern 
  out with factor level (instead of actualvalue). Remove dataframe factors.

This version doesn't work well with NA's in the dataframe.
}

