% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_dcb}
\alias{write_dcb}
\title{Write a TransCAD Dictionary File}
\usage{
write_dcb(df, file)
}
\arguments{
\item{df}{Data frame}

\item{file}{Name and path to the \code{.DCB} file.}
}
\description{
Write a TransCAD Dictionary File
}
