% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_df_to_binary}
\alias{write_df_to_binary}
\title{Write a binary TransCAD file from an R data frame.}
\usage{
write_df_to_binary(df, file, field_types)
}
\arguments{
\item{df}{List of variables (or a \code{data.frame}) that need to be
written out to TransCAD.}

\item{file}{A character string giving the output file name and path.}

\item{field_types}{A character vector giving the data types of each column.}
}
\description{
This is a C++ implementation written by Amar Sarvepalli and adapted for
Rcpp by Greg Macfarlane.
}
\details{
Currently this function only exports only 3 datatypes: Character,
  Integer and Double.
}
